+++ DvZ/src/me/andre111/dvz/monster/CustomMonster.java
@@ -36,7 +36,7 @@ public class CustomMonster {
 	private ArrayList<String> disabledDamage;
 	private int maxMana;
 	private int manaRegen;
-	private int startHealth;
+	private double startHealth;
 	private int startHunger;
 	private float startSat;
 	
@@ -199,10 +199,10 @@ public class CustomMonster {
 	public void setManaRegen(int manaRegen) {
 		this.manaRegen = manaRegen;
 	}
-	public int getStartHealth() {
+	public double getStartHealth() {
 		return startHealth;
 	}
-	public void setStartHealth(int startHealth) {
+	public void setStartHealth(double startHealth) {
 		this.startHealth = startHealth;
 	}
 	public int getStartHunger() {
